pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "academic-rag"
version = "1.0.0"
description = "Academic literature RAG system with hybrid search (BM25 + Vector + RRF), smart chunking, and Claude Code integration."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Academic RAG Contributors"},
]
keywords = ["rag", "academic", "retrieval", "literature", "search", "embedding", "chromadb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Indexing",
]

dependencies = [
    "openai>=1.0",
]

[project.optional-dependencies]
pdf = ["pymupdf>=1.23"]
docx = ["python-docx>=1.0"]
vector = ["chromadb>=0.4"]
local-embedding = ["sentence-transformers>=2.2"]
reranker = ["sentence-transformers>=2.2"]
all = [
    "pymupdf>=1.23",
    "python-docx>=1.0",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
academic-rag = "rag.cli:main"

[project.urls]
Homepage = "https://github.com/git-zyyang/academic-rag"
Repository = "https://github.com/git-zyyang/academic-rag"
Issues = "https://github.com/git-zyyang/academic-rag/issues"

[tool.setuptools.packages.find]
include = ["rag*"]
